# internal/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Portal data terbuka API — dataset, artikel, standar data, dan infografis.",
        "title": "Satudata API",
        "termsOfService": "https://satudata.go.id/terms",
        "contact": {
            "name": "Tim Satudata",
            "email": "info@satudata.go.id"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/admin/releases/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus release secara soft-delete (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Hapus release (soft-delete)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Release ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/protected/releases": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Membuat release baru (dataset, article, atau infographic)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Protected"
                ],
                "summary": "Buat release baru",
                "parameters": [
                    {
                        "description": "Data release",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateReleaseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ReleaseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            }
        },
        "/protected/releases/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Memperbarui data release yang sudah ada",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Protected"
                ],
                "summary": "Update release",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Release ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateReleaseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ReleaseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/public/releases": {
            "get": {
                "description": "Mengembalikan daftar releases dengan filter, pagination, dan sorting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Daftar releases",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by type: dataset, article, infographic",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status: draft, published, archived",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by year",
                        "name": "year",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search by title or description",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 10, max: 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field (default: created_at)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort direction: asc, desc (default: desc)",
                        "name": "sort_dir",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dto.ReleaseResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/public/releases/slug/{slug}": {
            "get": {
                "description": "Mengembalikan detail release berdasarkan slug (untuk URL SEO-friendly)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Detail release by Slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Release slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ReleaseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/public/releases/{id}": {
            "get": {
                "description": "Mengembalikan detail release berdasarkan ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Detail release by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Release ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/dto.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ReleaseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/dto.MetaData"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.ArticleMetadataResponse": {
            "type": "object",
            "properties": {
                "author_name": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "is_featured": {
                    "type": "boolean"
                },
                "reading_time_minutes": {
                    "type": "integer"
                }
            }
        },
        "dto.CreateReleaseRequest": {
            "type": "object",
            "required": [
                "release_type",
                "title",
                "year"
            ],
            "properties": {
                "author_name": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "content": {
                    "description": "Article-specific",
                    "type": "string"
                },
                "cover_image_url": {
                    "type": "string"
                },
                "data_period_end": {
                    "type": "string"
                },
                "data_period_start": {
                    "type": "string"
                },
                "data_source": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "excerpt": {
                    "type": "string"
                },
                "file_format": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_url": {
                    "description": "Dataset-specific",
                    "type": "string"
                },
                "is_featured": {
                    "type": "boolean"
                },
                "is_geospatial": {
                    "type": "boolean"
                },
                "release_type": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "update_frequency": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "dto.DatasetMetadataResponse": {
            "type": "object",
            "properties": {
                "column_count": {
                    "type": "integer"
                },
                "data_period_end": {
                    "type": "string"
                },
                "data_period_start": {
                    "type": "string"
                },
                "data_source": {
                    "type": "string"
                },
                "file_format": {
                    "type": "string"
                },
                "file_size": {
                    "type": "integer"
                },
                "file_url": {
                    "type": "string"
                },
                "is_geospatial": {
                    "type": "boolean"
                },
                "row_count": {
                    "type": "integer"
                },
                "update_frequency": {
                    "type": "string"
                }
            }
        },
...
        "dto.MetaData": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ReleaseResponse": {
            "type": "object",
            "properties": {
                "article_metadata": {
                    "$ref": "#/definitions/dto.ArticleMetadataResponse"
                },
                "cover_image_url": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "dataset_metadata": {
                    "$ref": "#/definitions/dto.DatasetMetadataResponse"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "release_type": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "view_count": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
        "dto.UpdateReleaseRequest": {
            "type": "object",
            "properties": {
                "cover_image_url": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Masukkan token dengan format: `Bearer \u003ctoken\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}